Extracted a writer class from runner
