+++ byoqm/runner/runner.py
@@ -1,19 +1,14 @@
 import logging
-import subprocess
-import time
-import csv
 import sys
-import pandas as pd
+
 
 import importlib.util
 import importlib.machinery
 
 from pathlib import Path
-from typing import Dict, List
+from typing import Dict
 
 from test.test_support import os
-from byoqm.metric.metric import Metric
-from byoqm.metric.result import Result
 
 from byoqm.qualitymodel.qualitymodel import QualityModel
 from byoqm.source_repository.source_repository import SourceRepository
@@ -26,17 +21,10 @@ class Runner:
         self,
         model_name: str,
         src_root: Path,
-        output_path: Path,
-        save_file: bool,
         language: str,
     ):
-        self._shortenPath = src_root
-        self._src_root: Path = src_root.resolve()
         self._model: QualityModel = self._load(model_name)
-        self._model_name: str = model_name
-        self._output_dir = output_path
-        self._save_file = save_file
-        self._source_repository = SourceRepository(self._src_root, language)
+        self._source_repository = SourceRepository(src_root.resolve(), language)
 
     def _load(self, model_name: str) -> QualityModel:
         """
@@ -60,17 +48,16 @@ class Runner:
             raise FileNotFoundError
         return path_to_model
 
-    def run(self) -> Path:
+    def run(self) -> Dict:
         """
         Returns a path to a .csv file that contains the metric and aggregation results
         as defined by the currently loaded model.
         """
-        results = self._run_aggregations()
-        if self._save_file:
-            self._gen_output_paths_if_not_exists()
-            output = self._write_to_csv(results)
-            return output
-        return None
+        try:
+            return self._run_aggregations()
+        except:
+            logging.error("failed to run aggregations")
+            return None
 
     def _run_aggregations(self) -> Dict:
         """
@@ -105,72 +92,3 @@ class Runner:
             results[metric] = module.metric.run()
         logging.info("Finished running metrics")
         return results
-
-    def _generate_violations_table(self, results: Dict, time: str):
-        """
-        _generate_violations_table generates a csv file in the output/violations/ folder containing
-        the locations of all found violations during parsing of the code base
-
-        _generate_violations_table given a dictionary of metrics will generate a csv file using
-        pandas. Because the dictionary may contain aggregation results that isn't of type Result,
-        and therefore doesn't contain locations of violations, we continue upon meeting an element
-        not of type Result.
-        """
-        list_of_violations = []
-        for _, result in results.items():
-            if type(result) is not Result:
-                continue
-            for location in result.get_violation_locations():
-                list_of_violations.extend([[result.metric, location]])
-        violations = pd.DataFrame(
-            list_of_violations, columns=["type", "file_start_end"]
-        )
-        file_location = Path(
-            self._output_dir / Path("violations") / Path(time + ".csv")
-        )
-        violations.to_csv(file_location)
-
-    def _write_to_csv(self, results: Dict):
-        """
-        _write_to_csv generates a csv file containing aggregation results and
-        underlying metrics
-
-        _write_to_csv will for each dictionary entry write either the associated
-        frequency of violations for a metric, or the aggregation result to a csv
-        file and output it in the frequencies folder
-
-        the naming naming format is YYYY-MM-DD_HH-MM-SS so that dashboard.py can
-        use it to display data chronologically
-        """
-        # See https://docs.python.org/3/library/time.html#time.strftime for table
-        # explaining formattng
-        # Format: YYYY-MM-DD_HH-MM-SS
-        logging.info("Writing to csv")
-        current_time: str = time.strftime("%Y-%m-%d_%H-%M-%S", time.gmtime())
-        file_name = Path(current_time + ".csv")
-
-        self._generate_violations_table(results, current_time)
-
-        file_location = self._output_dir / Path("frequencies") / file_name
-        with open(file_location, "w") as results_file:
-            writer = csv.writer(results_file)
-            writer.writerow(["metric", "value"])
-            for description, value in results.items():
-                frequency = value
-                if type(value) is Result:
-                    frequency = value.get_frequency()
-                writer.writerow([description, frequency])
-        logging.info("Finished writing frequencies to csv")
-        file_location = self._output_dir / Path("metadata") / file_name
-        with open(file_location, "w") as metadata_file:
-            writer = csv.writer(metadata_file)
-            writer.writerow([f"qualitymodel", "src_root"])
-            writer.writerow([self._model_name, self._shortenPath.__str__()])
-        logging.info("Finished writing metadata to csv")
-        return file_location
-
-    def _gen_output_paths_if_not_exists(self):
-        Path(self._output_dir).resolve().mkdir(exist_ok=True)
-        Path(self._output_dir / Path("violations")).resolve().mkdir(exist_ok=True)
-        Path(self._output_dir / Path("frequencies")).resolve().mkdir(exist_ok=True)
-        Path(self._output_dir / Path("metadata")).resolve().mkdir(exist_ok=True)
